Updated doc for Gantt changes. Renamed file and method (fixes #51)
